Included Rayleigh scattering in CLASS.
I followed the article arXiv:1307.8148. The main modifications are in
the thermodynamics module with the inclusion of a frequency
dependent Rayleigh scattering rate.

Then, I have added extra LOS sources, transfer functions and C_l's for
Rayleigh indices. Now, the Rayleigh C_l's are automatically produced
when the output is set to 'rCl'. The frequency is specified via the
setting 'rayleigh_frequency' (see rayleigh.ini parameter file).

Note that some details about the exact implementation of the LOS
sources are still unclear, and are likely to change in the future.