Reproduced P(k) from Wilkinson et al. 2013!
Implemented the DM-photon interaction in the perturbations.c module,
at the level of the differential system. The modifications are
concentrated at the end of the scalar part of 'perturb_derivs'
and coincide with Eqs. 5 and 6 of Wilkinson et al. 2013.
Note that the DM interactions are implemented only for Newtonian
gauge and with the radiation streaming approximation turned off.

Our results match with u=0, u=1e-2 and u=1e-4 match with those
in Fig. 5 of the paper, up to some numerical noise at high k's.
Next step is to check the Cl's, and