Obtained convergence in the kSZ power spectra.
Now P_q(k) converges with both the parallel and perpendicular kernels.
The normalisation factors and the redshift dependence seem to be correct,
as we match the dotted curve at z=9 in Fig. 1 of Park et al. (2013,
http://ukads.nottingham.ac.uk/abs/2013ApJ...769...93P).

We do not match however the dashed curves shown in Fig. 1 of Ma & Fry 2012.
I suspect they do something to their P_q(k), because I cannot see how
the parallel and perpendicular P_q(k) can be the same for large k's.

We still have a numerical issue that introduces a step at k~0.6 in the kSZ
spectra. That comes from the change in k-sampling in ppt->k at that scale.
It is fixed by setting k_per_decade_for_pk=1000, but it would be better to
understand why the step occurs. It is probably some error in the way I compute
the measure \Delta k for the trapezoidal integration.