Implemented slightly better kSZ integration technique.
The integration grid to obtain the kSZ power spectra (Eq. 6 and 7
of Ma & Fry 2002) used to be determined solely by the k-values in
ppt->k. Now for a given (k,k1) slice we make sure that both limits
of the triangular region of k2 are included (k2=|k-k1| and k2=k+k1). We compute
the integrand function there using interpolation.

However, the result is still wrong. We see this by comparison with
Ma & Fry and with a Mathematica notebook where we use interpolation
for the whole k2 (or mu1) domain. To have it right we should
include not only the limits of the triangular region, but also
enough points between them and the first element of ppt->k in
the range. The whole idea of using ppt->k to sample the k2 (i.e. mu1)
integral without resorting to interpolation was appealing but alas
not practical. In the next commit I will make sure that the k2 (i.e. mu1)
integral is solved by interpolating the integrand function rather than
using only the node points.