Fixed convolution to obtain the kSZ integrals in spectra.c.
We now take at least N points in each (k,k1) slice for the k2-integration. These are
important to sample the kernel which, for squeezed configurations, varies much more
rapidly than P(k2). For these extra points, the power spectrum could be interpolated,
but for now we just take it to be constant and equal to the nearest k-value to the left

Also, added 'lin_space' and 'log_space' functions to array.c, to
generate linearly and logarithmically spaced grids, respectively.