Added a_eq and k_eq variables to background structure.
The variable pba->a_eq is the value of the scale factor at the time
of matter-radiation equality.
The variable pba->k_eq is the comoving wavemode entering the horizon
at the time of matter-radiation equality, computed as 1/tau_eq, where
tau is the conformal time.
Both are computed using bispection.