Fixed l_max issue with lensing & bispectra.
When computing bispectra, we need the lensed C_l to be computed
up to the same l_max as the unlensed C_l. This can be achieved by increasing
the overall l_max by delta_l_max, as it is done in standard CLASS. However,
this becomes very computationally expensive when dealing with bispectra. Therefore,
we adopt a workaround whereby we increase l_max only for the spectra and lensing
modules, and not for the bispectra module. For details on how we do so, refer
to the compute_cls() function in the utility.c file.