Implemented a simple logging system.
I have defined a global variable `log_file` containing a pointer
to a text file. I have updated the following modules to support it:

 - bessel.c
 - bispectra.c
 - fisher.c

To print any message from CLASS or SONG to the log file and to screen,
use the new LOG(), printf_log() and printf_log_if() functions. With the
latter function, more information will be printed to the log file than
what is printed to screen, so that inspecting the log file is always
informative.