Introduced the TCA1 scheme from http://arxiv.org/abs/1012.0546
This scheme from Pitrou 2011 is simpler than those implemented
in CLASS because it does not involve derivatives of the collision
time tau_c at first order. So far it works fine, but I have
to check the coefficients for the expression for the shear, because
Pitrou uses a peculiar type of variables.