Evolver now calls print function more often for SONG.
In SONG, we call the print_variables() function for every time step AND
for every time in the time sampling array. The name of the function
actually called is perturb2_save_perturbations(). In this way, SONG is
able to print output for all values of tau AND of k3.

Also moved the binary structures and functions to the new binary.c and
binary.h files.