IN PROGRESS
Trying to implement arbitrary l3 sampling in bispectra_lensing(),

First step I tried is to make index_l3 -> l3 in
bispectra_lensing_convolution(), but I soon realised that for this
to work, 3D interpolation of the bispectrum is needed to make the
symmetrisation work. Just as for making arbitrary l3 sampling for
all bispectra and not only the lensing correction.

So I guess that the next step would be to implement 3D bispectra
computation.

See `git diff HEAD^1` to see what has changed.